requirements.txt
django>=5.2.0,<6.0.0
djangorestframework>=3.16.0,<4.0.0
django-filter>=25.1,<26.0
psycopg2-binary>=2.9.9,<3.0.0
djangorestframework-simplejwt>=5.5.0,<6.0.0
drf-yasg>=1.21.7,<2.0.0
python-dotenv>=1.1.0,<2.0.0
pytest>=8.3.0,<9.0.0
pytest-django>=4.11.0,<5.0.0
gunicorn>=21.2.0,<22.0.0
coreapi>=2.3.3,<3.0.0
swagger-ui-bundle>=0.0.9
pyyaml>=6.0.1,<7.0.0
uritemplate>=4.1.1,<5.0.0 